requirements.txt
pytorch
opencv
torchvision
